Move EnsureCapacity to its own file
